actions/execute_atlas_pipeline/scripts/requirements.txt
requests
pyyaml
qubership-pipelines-common-library